#103 used a wrapper class instead of interface
